+++ category.h
@@ -1,18 +1,18 @@
 #include "name.h"
 
 typedef struct _category {
-   float split;
+   double split;
    Name *head, *tail;
    struct _category *left, *right;
 
-   float top;
+   double top;
    struct _category *next, *prev;
 } Category;
 
 Category *create_category(void);
-Category *get_category(Category *cat, float x);
+Category *get_category(Category *cat, double x);
 Category *left_most(Category *cat);
-void inner_split(Category *cat, float x, float y);
+void inner_split(Category *cat, double x, double y);
 int peek(Category *cat);
 void push(Category *cat, int value, int time_stamp);
 void enqueue(Category *cat, int value, int time_stamp);
